refactor(hir): extract hir-def crate
