+++ crates/hir-def/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "hir"
+name = "hir-def"
 version = "0.0.0"
 edition.workspace = true
 
 [dependencies]
-hashbrown = { version = "0.12.3", features = ["inline-more"], default-features = false }
 base-db.workspace = true
-hir-def.workspace = true
+hashbrown = { version = "0.12.3", features = ["inline-more"], default-features = false }
 itertools.workspace = true
 la-arena.workspace = true
 preproc-expand.workspace = true
