+++ crates/hir-def/src/hir_def/declaration.rs
