+++ crates/hir-def/src/hir_def/module/generate.rs
