+++ crates/hir-def/src/symbol.rs
