+++ crates/hir/src/lib.rs
@@ -1,19 +1,11 @@
-#![feature(decl_macro)]
-
+pub use ::hir_def::{
+    container, def_id, has_source, hir_def, pathres, region_tree, scope, source_map, symbol,
+};
 pub use base_db::{self, impl_intern_key, impl_intern_lookup};
 pub use preproc_expand::{self, compilation_plan, file, macro_file, preproc};
-pub mod container;
 pub mod db;
-pub mod def_id;
 pub mod display;
-pub mod has_source;
-pub mod hir_def;
-pub mod pathres;
-pub mod region_tree;
-pub mod scope;
 pub mod semantics;
-pub mod source_map;
-pub mod symbol;
 mod ty_db;
 pub mod type_infer;
